Podpisanej oferty nie mozna usunac
Oferta z podpisem to zapis tego, co klient zaakceptowal, a event
wskazywal na nia zwykla kolumna bez klucza obcego - usuniecie
oferty zostawialo event z wiszacym odwolaniem i psulo link
w widoku eventu.

Jedyna droga usuniecia takiej oferty jest teraz usuniecie eventu,
ktory ja utworzyl.

+++ src/main/java/com/eventsApp/offer/OfferService.java
@@ -45,6 +45,10 @@ public class OfferService {
 
     public void delete(int id) {
         Offer offer = getOwnedOffer(id);
+        if (offer.getStatus() == OfferStatus.SIGNED) {
+            throw new EventApiException(
+                    "Signed offer can not be deleted", HttpStatus.CONFLICT);
+        }
         offerRepository.delete(offer);
     }
